pyproject.toml
[project]
name = "sam3-pursuit"
version = "1.0.0"
description = "Fursuit character recognition using SAM3 and DINOv2"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "ultralytics>=8.3.0",
    "transformers>=4.40.0",
    "faiss-cpu>=1.7.4",
    "pillow>=10.0",
    "numpy>=1.24",
    "requests>=2.31.0",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sam3_pursuit*"]
